requirements.txt
streamlit
torch
torchvision
transformers
vaderSentiment
demoji
pillow
numpy
pandas
shap
opencv-python
matplotlib